Moved FeaturedMemberController Components into the FeaturedMemberController directory
