+++ src/CharityDashboard/components/FeaturedMemberController/FeaturedMemberPreview/FeaturedMemberPreview.scss
