+++ src/CharityDashboard/components/FeaturedMemberController/FeaturedMemberSelector/FeaturedMemberSelector.scss
@@ -11,6 +11,7 @@
     color: white;
     background: #444;
   }
+  
   .del-div{
     height: 19px;
   }
